+++ src/components/Icon/Icon.tsx
@@ -1,5 +1,3 @@
-"use server";
-
 import React from 'react';
 import {IconProps} from "@phosphor-icons/react";
