+++ src/components/index.ts
@@ -1,3 +1,4 @@
 export {Button} from './Button';
 export {Icon} from './Icon';
 export {Label} from './Label';
+export {Input} from './Input';
