+++ src/components/theme/constants.ts
@@ -1,4 +1,1 @@
-export const LIGHT_THEME = 'light';
-export const DARK_THEME = 'dark';
-
-export const DEFAULT_THEME = LIGHT_THEME;
+export const LightTheme = 'light', DarkTheme = 'dark', DefaultTheme = LightTheme;
