+++ src/components/theme/useTheme.tsx
@@ -1,4 +1,5 @@
-import * as React from "react";
+import React from "react";
+
 import {ThemeContext} from "./ThemeContext";
 
 export function useTheme() {
